Calculate total delivery fees
Return the sum of regional base fees and extra fees based on weather conditions.

+++ src/main/java/com/fujitsu/delivery/controller/DeliveryController.java
@@ -4,6 +4,8 @@ import com.fujitsu.delivery.service.DeliveryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
+
 @RestController
 @RequestMapping("/api/delivery")
 public class DeliveryController {
@@ -12,8 +14,8 @@ public class DeliveryController {
     private DeliveryService deliveryService;
 
     @GetMapping(value = "getDeliveryFees")
-    public int getDeliveryFees(@RequestParam(value = "city") String city,
-                               @RequestParam(value = "vehicleType") String vehicleType) {
+    public BigDecimal getDeliveryFees(@RequestParam(value = "city") String city,
+                                      @RequestParam(value = "vehicleType") String vehicleType) {
         return deliveryService.getDeliveryFees(city, vehicleType);
     }
 }
